backend/src/services/file_analysis.py
from flask import Blueprint, request, jsonify
import openai

file_analysis_bp = Blueprint('file_analysis', __name__)

@file_analysis_bp.route('/analyze', methods=['POST'])
def analyze_file():
    file = request.files.get('file')
    if not file:
        return jsonify({'error': 'No file provided'}), 400

    # Here you would implement the logic to analyze the file
    # For example, if it's an image, you might use GPT-4 Vision
    # This is a placeholder for the actual analysis logic
    analysis_result = perform_analysis(file)

    return jsonify({'result': analysis_result}), 200

def perform_analysis(file):
    # Placeholder function for file analysis logic
    # You can integrate with GPT-4 Vision or other AI tools here
    return "Analysis result for the uploaded file."