backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
openai
motor
python-dotenv
apscheduler